+++ lib/catan.rb
@@ -11,3 +11,5 @@ Dir[File.join(__dir__, 'catan', 'helpers', '*.rb')].each { |file| require file }
 
 require File.join(__dir__, 'catan', 'interactors', 'interactor.rb')
 Dir[File.join(__dir__, 'catan', 'interactors', '*.rb')].each { |file| require file }
+
+Dir[File.join(__dir__, 'catan', 'serializers', '*.rb')].each { |file| require file }